Rename predefined answers module to static answers
